# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "description": "API для керування новинами з авторизацією через Sanctum",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "description": "Отримати профіль поточного користувача",
                "operationId": "b70905c3804ac0196a066fc4256bc24c",
                "responses": {
                    "200": {
                        "description": "Профіль користувача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Помилка сервера"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Profile"
                ],
                "description": "Оновити профіль користувача (використовуйте POST з _method=PUT)",
                "operationId": "ae932c2c21ae5c4a890c7222d4f2c9b2",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Іван Оновлений"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "newemail@example.com"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Новий опис профілю"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Профіль успішно оновлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Помилка сервера"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile/avatar": {
            "post": {
                "tags": [
                    "User Profile"
                ],
                "description": "Завантажити або оновити аватар користувача",
                "operationId": "7f77c3e0c05aba840507ff89038120a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "description": "Зображення аватара (jpg, png, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Аватар успішно завантажено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "avatar": {
                                            "type": "string",
                                            "example": "avatars/avatar123.jpg"
                                        },
                                        "avatar_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/storage/avatars/avatar123.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Помилка сервера"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Profile"
                ],
                "description": "Видалити аватар користувача",
                "operationId": "f3df66239d0e4101e19a9fea74246f03",
                "responses": {
                    "200": {
                        "description": "Аватар успішно видалено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Аватар успішно видалено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Аватар не знайдено"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "Модель користувача",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Іван Петренко"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Опис користувача",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "example": "avatars/avatar.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Введіть токен отриманий після логіну/реєстрації",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}